+++ src/Samsara.Sdk/Serialization/SamsaraJsonContext.cs
@@ -111,6 +111,8 @@ using Samsara.Sdk.Pagination;
 [JsonSerializable(typeof(DriverSafetyScore))]
 [JsonSerializable(typeof(TagSafetyScore))]
 [JsonSerializable(typeof(TagGroupSafetyScore))]
+[JsonSerializable(typeof(SafetyScoreBehavior))]
+[JsonSerializable(typeof(SafetyScoreSpeeding))]
 [JsonSerializable(typeof(TimeRange))]
 // Routes
 [JsonSerializable(typeof(Route))]